backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
Authlib==1.6.9
cffi==2.0.0
click==8.3.2
colorama==0.4.6
cryptography==46.0.7
fastapi==0.135.3
h11==0.16.0
idna==3.11
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
python-dotenv==1.2.2
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
sqlalchemy==2.0.38
psycopg2-binary==2.9.10
redis==5.3.1
fastapi-cache2==0.2.2
orjson==3.10.15
boto3==1.37.16
botocore==1.37.16
email-validator==2.2.0
gunicorn==23.0.0
python-multipart==0.0.20
starlette==1.0.0
itsdangerous==2.2.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.44.0
